fix: divide device & device status class to fix cache issue
